feat(genome): externalize prompts and heuristics
