+++ genome/heuristics/failure-strategy.ts
@@ -1,0 +1,11 @@
+import {
+  parseFailureStrategyHeuristicConfig,
+  type FailureStrategyHeuristicConfig
+} from '@evolvo/core/heuristic-config';
+
+export const failureStrategyHeuristics: FailureStrategyHeuristicConfig =
+  parseFailureStrategyHeuristicConfig({
+    directFixMaxRecurrence: 1,
+    mutationFirstLowConfidenceThreshold: 45,
+    mutationFirstMinRecurrence: 2
+  });
