+++ genome/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "@evolvo/typescript-config/library.json",
+  "include": ["**/*.ts"]
+}
