+++ packages/scoring/package.json
@@ -15,6 +15,7 @@
     "typecheck": "tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
+    "@evolvo/genome": "workspace:*",
     "@evolvo/schemas": "workspace:*"
   },
   "devDependencies": {
