+++ packages/scoring/src/issue-priority.ts
@@ -1,3 +1,4 @@
+import { issuePriorityHeuristics } from '@evolvo/genome/heuristics/issue-priority';
 import type {
   FailureCategory,
   MutationState,
@@ -24,56 +25,35 @@ function clampScore(score: number): number {
 }
 
 function resolveRiskWeight(riskLevel: RiskLevel | undefined): number {
-  switch (riskLevel) {
-    case 'systemic':
-      return 24;
-    case 'high':
-      return 16;
-    case 'medium':
-      return 8;
-    case 'low':
-    default:
-      return 0;
-  }
+  return issuePriorityHeuristics.riskWeights[riskLevel ?? 'low'];
 }
 
 function resolveMutationWeight(mutationState: MutationState | undefined): number {
-  switch (mutationState) {
-    case 'validated':
-      return 18;
-    case 'selected':
-    case 'in-progress':
-      return 12;
-    case 'proposed':
-      return 6;
-    default:
-      return 0;
+  if (!mutationState) {
+    return 0;
   }
+
+  return issuePriorityHeuristics.mutationStateWeights[mutationState];
 }
 
 function resolveCategoryWeight(category: FailureCategory | undefined): number {
-  switch (category) {
-    case 'model-quality-issue':
-    case 'dependency-configuration-issue':
-      return 12;
-    case 'smoke-e2e-failure':
-    case 'runtime-failure':
-      return 8;
-    case 'requirement-misunderstanding':
-      return 4;
-    default:
-      return 0;
+  if (!category) {
+    return 0;
   }
+
+  return issuePriorityHeuristics.categoryWeights[category];
 }
 
 export function buildIssuePriorityScore(input: BuildIssuePriorityInput): number {
   return clampScore(
-    (input.basePriorityScore ?? 40) +
-      Math.max(0, (input.recurrenceCount ?? 1) - 1) * 12 +
+    (input.basePriorityScore ??
+      issuePriorityHeuristics.basePriorityScoreDefault) +
+      Math.max(0, (input.recurrenceCount ?? 1) - 1) *
+        issuePriorityHeuristics.recurrenceBonusPerRepeat +
       resolveRiskWeight(input.riskLevel) +
       resolveMutationWeight(input.mutationState) +
       resolveCategoryWeight(input.category) +
-      (input.isSystemic ? 15 : 0)
+      (input.isSystemic ? issuePriorityHeuristics.systemicBonus : 0)
   );
 }
